==> 3_validity/simplexes/3_merge_annots.ipynb <==
Merges the manual annotations in `infiles/simplex_filtered2_annotated.csv` with `outfiles/simplex_filtered2.csv` to create `outfiles/simplex_filtered3.csv`, a file that should contain mainly simplexes, their POS, and their frequency in DECOW16B.

import pandas as pd

filt2 = pd.read_csv('outfiles/simplex_filtered2.csv')
annot = pd.read_csv('infiles/simplex_filtered2_annotated.csv')

# Merge the two dataframes, adding the 'simplex' column onto filt2.
filt3 = filt2.merge(annot)

# Only keep those rows in which simplex == 1, remove now-redundant col, and save to csv.
filt3[ filt3['simplex'] == 1 ].drop(columns=['simplex']).to_csv('outfiles/simplex_filtered3.csv', index=False)
